drivers: serial: ns16550: Support for UART interface on JZ4740
The UART interface on the JZ4740 SoC is not strictly compatible with
the 16550 implementation. The baud rate formula on the programming
manual was wrong. Presumably on older versions of kernel and u-boot
the calculation does not round up, so it was working properly.

The baud rate should be:

baud = fEXT / (16 * (Divisor + 1))

(Tested on actual device)